绑定机器人 上传数据 状态控制

+++ born/src/main/java/com/zbwang/face/dao/IUserDao.java
@@ -3,11 +3,9 @@ package com.zbwang.face.dao;
 import com.zbwang.face.domain.User;
 
 public interface IUserDao {
-
-	Integer insertUser(User user);
+	Long insertUser(User user);
 
 	User getUser(String userName);
 
 	User getUser(Integer userId);
-
 }
