+++ born/src/main/java/com/zbwang/face/service/IUserService.java
@@ -3,11 +3,9 @@ package com.zbwang.face.service;
 import com.zbwang.face.domain.User;
 
 public interface IUserService {
-
-	Integer insertUser(User user);
+	Long insertUser(User user);
 
 	User getUser(String userName);
 
 	User getUser(Integer userId);
-
 }
